fix: fix small style issue and form population issue
