+++ src/modules/75-cd/components/PipelineSteps/Common/Terraform/Editview/TerraformVarfile.module.scss
@@ -175,6 +175,7 @@
   }
 
   .formContainerStepOne {
+    margin-top: 20px !important;
     height: 100%;
     min-height: 280px;
   }
